Update: Added: refactoring workflow
