+++ tests_ui/conftest.py
@@ -7,7 +7,7 @@ from selenium import webdriver
 from selenium.webdriver.chrome.options import Options
 
 from tests_ui.links import TestDataLinks
-from tests_ui.pages.login_page import LoginPage
+from tests_ui.login_page import LoginPage
 
 
 @pytest.fixture()
