+++ tests_ui/login_page.py
@@ -5,7 +5,7 @@ import allure
 from dotenv import load_dotenv
 
 from tests_ui.login_page_locators import Locators
-from tests_ui.pages.base_page import BasePage
+from tests_ui.base_page import BasePage
 
 load_dotenv()
